@@ -9,7 +9,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/wix/react-native-ui-lib"
+    "url": "git+https://github.com/wix/react-native-ui-lib.git"
   },
   "scripts": {
     "release": "node ./scripts/release.js",
@@ -20,8 +20,8 @@
     "shell-utils": "^1.0.10"
   },
   "peerDependencies": {
-    "react": "15.4.2",
-    "react-native": "0.38.0",
+    "react": "^16.0.0",
+    "react-native": "^0.51.0",
     "react-native-ui-lib": "*",
     "semver": "^5.5.0"
   }
