@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import {ScrollView, StyleSheet} from 'react-native';
 import {Colors, View, Badge, Button, Text} from 'react-native-ui-lib';//eslint-disable-line
-
 const BadgesSpace = 30;
 const plusIcon = require('../../assets/icons/chevronUp.png');
 const minusIcon = require('../../assets/icons/chevronDown.png');
@@ -27,17 +26,17 @@ export default class BadgesScreen extends Component {
   render() {
     return (
       <ScrollView contentContainerStyle={styles.container}>
+        <Text text50 row center marginB-15> Badges </Text>
         <View row center>
-          <View center marginR-20>
-            <Badge label={'1'} containerStyle={{marginTop: BadgesSpace}} backgroundColor={Colors.blue30}/>
+          <View center marginR-10 marginL-10>
+            <Badge label={'1'} backgroundColor={Colors.blue30}/>
             <Badge label={'1'} containerStyle={{marginTop: BadgesSpace}} backgroundColor={Colors.blue30} borderWidth={3} borderColor={Colors.blue50}/>
             <Badge size="small" label={'1'} containerStyle={{marginTop: BadgesSpace}} backgroundColor={Colors.blue30}/>
           </View>
 
-          <View center marginR-20>
+          <View center marginR-10 marginL-10>
             <Badge
               label={this.state.value.toString()}
-              containerStyle={{marginTop: BadgesSpace}}
               backgroundColor={Colors.red30}
             />
             <Badge
@@ -54,14 +53,14 @@ export default class BadgesScreen extends Component {
             />
           </View>
 
-          <View center marginR-20>
-            <Badge label={'99+'} containerStyle={{marginTop: BadgesSpace}}/>
+          <View center marginR-10 marginL-10>
+            <Badge label={'99+'}/>
             <Badge label={'99+'} containerStyle={{marginTop: BadgesSpace}} borderWidth={2} borderColor={Colors.white}/>
             <Badge size="small" label={'99+'} containerStyle={{marginTop: BadgesSpace}}/>
           </View>
         </View>
 
-        <View row marginT-50 marginB-15>
+        <View row marginT-20 marginB-15>
           <Button
             bg-dark60 style={{width: 40, height: 40, borderWidth: 1, marginRight: 15}}
             iconSource={minusIcon}
@@ -78,6 +77,19 @@ export default class BadgesScreen extends Component {
           />
         </View>
         <Text center>Press buttons to change red badge value by 1. Long press to change it by 10.</Text>
+
+        <Text text50 row center marginT-50 marginB-15> Pimple Badges</Text>
+        <Text text80 row center marginB-15>size={'{\'pimpleSmall\'}'}</Text>
+        <View row center>
+          <Badge containerStyle={{marginRight: BadgesSpace}} backgroundColor={Colors.red30}/>
+          <Badge containerStyle={{marginLeft: BadgesSpace}} backgroundColor={Colors.blue30}/>
+        </View>
+
+        <Text text80 row center marginT-20 marginB-15>size={'{\'pimpleBig\'}'}</Text>
+        <View row center>
+          <Badge containerStyle={{marginRight: BadgesSpace}} backgroundColor={Colors.red30} size={'pimpleBig'}/>
+          <Badge containerStyle={{marginLeft: BadgesSpace}} backgroundColor={Colors.blue30} size={'pimpleBig'}/>
+        </View>
       </ScrollView>
     );
   }
